Commiting the current prototype scene I have. Also commiting the beat map player scripts I have so far.
Beat Map Player Script: Responsible for taking a beat map and playing its audio track while also generating beat sync data maps in realtime for the user to follow. The Beat Map Player will ensure that the audio track its playing is looping seamlessly until told otherwise. It can also be used to fade out the track it's currently playing. Songs are played with an intro -> loop -> outro in that order, the loop is playing infinitely until the script is told to stop looping. The trimming of a loop can be setup in the BeatMap scriptable object (whether or not an intro / outro should be used at all etc).

Beat Map Player Manager: Responsible for managing the currently played song between two beat map players. It will essentially be like the DJ of the game (play out Track A -> play in Track B, play out Track B -> play in Track A). The Beat Map Player Manager also tells beat map player scripts how much time the player should be warned about a beat before it's audible.

Adjusted the BeatMap class so it can split songs by an intro, loop, outro and save the audio clips to the disc instead of splitting audio clips during runtime.

Created a SaveWave file, it basically takes an audio clip and saves it to the disc (used by the BeatMap class)

Added some sample beats to play around with.

Added an audio mixer for songs.