Created a BeatSyncReceiver script
The BeatSyncReceiver class is responsible for listening for incoming beats emitted by the BeatMapPlayer classes. The BeatSyncReceiver class will be responsible for handling the logic for our game's beats. BeatMapPlayers emit beats early to the BeatSyncReceiver based on the "Beat Map Preview Time" of the Beat Map Player Manager in the scene. We may change the duration of this time for different difficulties or faster songs (the shorter the preview time, the less time the player will be warned about an upcoming beat).