Made a prototype scene for the UI of the beat maps. This is the first version, I'll be working on another version of the UI closer to the drawings that GummiGamer made.
Open the "Prototype Scene - UI Prototyping" scene in the "Scenes" folder. Play the game! If you want to press the "Turn Off Loop" button in the UI menu (for testing), you have to toggle off "Lock Mouse Cursor To Screen On Start" in the PlayerController script on the Player game object before you play the scene.

I had to download a Unity package for inverse image masking for the beat-map sample icons, the dependency package may be found in the "Imported Packages" folder. The package basically lets you create "reverse masks" for images.

The Unity package complies with the game jam rules of being free and open-source.

Link to the package: https://github.com/mob-sakai/UnmaskForUGUI#demo