Fixed some Unity Package import conflicts that happened across our commits.
The following imported packages now work fine: Post Processing, Pro Grids, Pro Builder, and Quci Search.

The Player prefab might be broken because the gun appears to be missing from the FPS hands. In any case, we can always pull a previous version of the prefab if this wasn't intended. :)