Modified the BeatMap scriptable objects to generate beat timestamps based on a BPM isntead of a preset list. Simply fill the beatmaps with a songs "BPM" and the timestamp of the first beat in the song and (optionally) the last beat in the song. The BeaMap then auto-generates a list of beat timestamps based on your settings.
BeatMapPlayerManager changes: Now has a bool value in the inspector that let's you toggle between syncing beats to preset beat timestamp lists or BPM auto-generated lists (simply check the "UseBPM" option here if you want the beatmap players to use the auto-generated BPM map). It may also be useful for beat preview time to be limited to the BPM of the actual song when using a BPM timestamp list, you may use the "Use BPM As Preview Time" option to automatically set the beat preview time to the bpm of your song.