Experimenting with different beat map UI indicator. It's been surprisingly challenging to make a beatmap indicator for the player while providing the spacing and freedom to focus on shooting.
Please check out the "Prototype UI" scenes.

Version 1: The original beatmap circles scrolling across the screen into the crosshair.

Version 2: The beatmap meter (the least invasive but doesn't permit consecutive beats. This is because the meter needs time to unfill before it can be ready to be filled up again)

Version 3: Similar to version 1 but the beatmaps are small lines instead of circles. Also changed the crosshair to a line to compliment the beatmap sprites.

Version 4: The beatmaps move along in a circular track below the crosshair. When the beatmap is directly under the crosshair, the player may execute an input.

We may have to mix and match the different beatmap UI styles for perfect blend between any of the UI. Each one has strengths and weaknesses. Will have to let the team test the different versions for feedback to decide which one to go with.