The mech enemy script is fully working! It can shoot projectiles at a given target, you can adjust the mech's attack settings in the "MechAttackScript".
Also added a "Bullet" prefab and script. Feel free to use it on any enemy prefab. All you have to do is instantiate the "Bullet" prefab, use the "Bullet" script and call "ProjectBullet", you'll  have to give it a target position, the damage it will deal if it hits something, and the layermask applied to it when it checks for collisions (for example you don't want enemies to be able to shoot and deal damage to other enemies). You may also optionally give it a projectile speed.

The bullet projectile still needs to handle the actual dealing damage when it colliding with the player, however I'll have to review how our health system works so I can use it in our bullet script.

Note: The bullet prefab uses the model found in the sci-fi weapons pack we downloaded (Link: http://devassets.com/assets/sci-fi-weapons/)