- Added a health pack prefab that the player can pickup for health, feel free to adjust the amount of health it rewards -  Added gift health funcationality to the player health script (maybe we can use this for combo bonuses too?) Health pack model link: https://sketchfab.com/3d-models/health-pack-baab802b496548368dabe5a95ccf3c54
Note: Note sure if we're even going to need this health pack prefab. If the game is too easy with the health packs, then let's cut it from our project.  Also feel free to use another model for the health pack instead of the one I used here, not really liking it anyway. It could also use some kind of particle effect to get the player's attention, it's pretty boring as it is.